Add basic dummy plugin support for .omwaddon.esp and .omwscripts.esp
If you create a file alongside your omwaddon/omscripts with .esp added at the end, the exporter will add the relevant file at that point in the load order without the .esp extension. This preserves addons/scripts position in the load order according to where you have the dummy plugin in MO2.

This does not yet create the dummy esp files for you.